Fix the ROS/Gtk threading assumptions
The ROS threading model requires calling init_node from the
main thread (because it may setup signal handlers).
Each subscriber (and publisher, not used here)
is assigned a thread and callbacks happen in that thread.
It is not even necessary to call rospy.spin() in the subscriber
case (with signals disabled) however one should call rospy.spin
for future compatibility. The conclusion of this is that one should
 * call init_node from the main thread
 * feel free to call rospy.spin in a non-main thread
 * remember data callbacks occur in non-main threads

Gtk is thread-aware but not thread safe. You can not update
UI state from non-main threads, but Gtk programs can use threading
providing the appropriate thread_init functions are called before
constructing anything. This means that if you recieve data in a non-main
thread an wish to update the UI there are two recommended patterns
 * use the GLib.idle_add API to schedule a function to be called from the
   main (gtk thread)
 * take the gtk lock (not recommneded because of problems on windows)
 * Use GObject signals (which are list of function callbacks) and 'emit'
   those signals on idle - this is essentially the first option, but fits
   more with the common model/view+controller split of Gtk applications
 * serialise UI updates to occur at regular intervals from the main thread

I took the last approach because I expect data to be arriving faster than
we can show it (i.e. sync to vblank limits the max framerate to 60Hz or so,
and we really can see much benefit in FPS > 30 anyway).

This means that I
 * schedule a timer to fire at the interval I want to update the UI
 * get data in a random subscriber thread, take a lock, and store it
 * get the timer callback, take the lock and update the UI